Refactor ContentCoordinator to simplify lookup
Single parameter functions that take and return optionals are a code smell – it usually means we should’ve dealt with the optionality sooner. For the case of `ContentCoordinator`, it would probably make sense to make `siteID` non-optional, and require callers not to bother using the method unless they already know they have a `siteID` present.

That said, the “right” solution is probably to require a non-optional `Blog` object instead and an assertion that we’re operating on the main thread (which would be required anyway since we’re dealing with UINavigationViewController pushes).

That refactor is outside the scope of this PR, but _may_ make sense when threading the Core Data Stack instance through this object.